# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Key Service",
    "description": "Manages API keys and BYOK (Bring Your Own Key) credentials for organizations. Handles key generation, validation, encryption, and secure storage.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "serviceKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service-to-service key (KEY_SERVICE_API_KEY)"
      }
    },
    "schemas": {
...
      "ListUserAuthKeysResponse": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserAuthKeyItem"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
      "UserAuthKeyItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "keyPrefix": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "orgId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "keyPrefix",
          "name",
          "orgId",
          "userId",
          "createdBy",
          "createdAt",
          "lastUsedAt"
        ]
      },
      "CreateUserAuthKeyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "orgId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "orgId",
          "userId",
          "createdBy",
          "createdAt"
        ]
      },
      "CreateUserAuthKeyRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "createdBy": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "userId",
          "createdBy",
          "name"
        ]
      },
...
      "ListOrgKeysResponse": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgKeyItem"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
      "OrgKeyItem": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateOrgKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "message"
        ]
      },
      "CreateOrgKeyRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "minLength": 1
          },
          "apiKey": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "provider",
          "apiKey"
        ]
      },
...
      "CreatePlatformKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "message"
        ]
      },
      "CreatePlatformKeyRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "minLength": 1
          },
          "apiKey": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "provider",
          "apiKey"
        ]
      },
      "ListPlatformKeysResponse": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformKeyItem"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
      "PlatformKeyItem": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ProviderRequirementsResponse": {
        "type": "object",
        "properties": {
          "requirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderRequirementItem"
            }
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "requirements",
          "providers"
        ]
      },
      "ProviderRequirementItem": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          }
        },
        "required": [
          "service",
          "method",
          "path",
          "provider"
        ]
      },
      "ProviderRequirementsRequest": {
        "type": "object",
        "properties": {
          "endpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Endpoint"
            },
            "minItems": 1
          }
        },
        "required": [
          "endpoints"
        ]
      },
      "Endpoint": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string",
            "minLength": 1
          },
          "method": {
            "type": "string",
            "minLength": 1
          },
          "path": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "service",
          "method",
          "path"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api-keys": {
      "get": {
        "summary": "List user auth keys for an org",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "userId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of user auth keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUserAuthKeysResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user auth key",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserAuthKeyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User auth key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserAuthKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/keys": {
      "get": {
        "summary": "List org keys for an org",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of org keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrgKeysResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add or update an org key",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgKeyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Org key saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrgKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/platform-keys": {
      "post": {
        "summary": "Add or update a platform key",
        "description": "Upsert a platform-level API key. Platform keys are global — not tied to any org. Typically registered once at deployment. No identity headers needed.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlatformKeyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Platform key saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePlatformKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      },
      "get": {
        "summary": "List all platform keys",
        "description": "No identity headers needed.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of platform keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPlatformKeysResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/provider-requirements": {
      "post": {
        "summary": "Query which providers are needed for a set of endpoints",
        "description": "Given a list of service endpoints (service + method + path), returns which third-party providers each endpoint has been observed requesting. Used by workflow-service to determine which keys are needed before execution. No identity headers needed.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderRequirementsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider requirements for the given endpoints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderRequirementsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
  }
}