# src/index.ts
// Note: Sentry is initialized via --import flag in package.json/Dockerfile
import * as Sentry from "@sentry/node";
import express from "express";
import cors from "cors";
import { readFileSync, existsSync } from "fs";
import { fileURLToPath } from "url";
import { dirname, join } from "path";
import { migrate } from "drizzle-orm/postgres-js/migrator";
import { db } from "./db/index.js";
import healthRoutes from "./routes/health.js";
import validateRoutes from "./routes/validate.js";
import keysRoutes from "./routes/keys.js";
import apiKeysRoutes from "./routes/api-keys.js";
import platformKeysRoutes from "./routes/platform-keys.js";
import providerRequirementsRoutes from "./routes/provider-requirements.js";
import { serviceKeyAuth, requireIdentityHeaders } from "./middleware/auth.js";

const __filename = fileURLToPath(import.meta.url);
const __dirname = dirname(__filename);

const app = express();
const PORT = process.env.PORT || 3001;

app.use(cors());
app.use(express.json());

// OpenAPI spec (public)
const openapiPath = join(__dirname, "..", "openapi.json");
app.get("/openapi.json", (_req, res) => {
  if (existsSync(openapiPath)) {
    res.json(JSON.parse(readFileSync(openapiPath, "utf-8")));
  } else {
    res.status(404).json({ error: "OpenAPI spec not generated. Run: pnpm generate:openapi" });
  }
});

// Health check (public)
app.use(healthRoutes);

// API key validation (service-to-service, authenticated via X-API-Key)
app.use(serviceKeyAuth, validateRoutes);

// Key management endpoints (resolve, preferences, org key CRUD)
app.use("/keys", serviceKeyAuth, requireIdentityHeaders, keysRoutes);

// User auth key CRUD
app.use("/api-keys", serviceKeyAuth, requireIdentityHeaders, apiKeysRoutes);

// Platform keys — no identity headers needed (cold-start bootstrap)
app.use("/platform-keys", serviceKeyAuth, platformKeysRoutes);

// Provider requirements — no identity headers needed
app.use("/provider-requirements", serviceKeyAuth, providerRequirementsRoutes);

// 404
app.use((req, res) => {
  res.status(404).json({ error: "Not found" });
});

// Sentry error handler must be before any other error middleware
Sentry.setupExpressErrorHandler(app);

// Fallback error handler
app.use((err: Error, req: express.Request, res: express.Response, next: express.NextFunction) => {
  console.error("Error:", err);
  res.status(500).json({ error: "Internal server error" });
});

// Only start server if not in test environment
if (process.env.NODE_ENV !== "test") {
  if (!process.env.KEY_SERVICE_API_KEY?.trim()) {
    console.error("[KEY SERVICE] FATAL: KEY_SERVICE_API_KEY env var is missing or empty. Cannot start.");
    process.exit(1);
  }

  migrate(db, { migrationsFolder: "./drizzle" })
    .then(() => {
      console.log("Migrations complete");
      app.listen(Number(PORT), "::", () => {
        console.log(`Key service running on port ${PORT}`);
      });
    })
    .catch((err) => {
      console.error("Migration failed:", err);
      process.exit(1);
    });
}

export default app;